+++ src/Attribute/Data/DataAttributeInterface.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Hydrator\AttributeInfrastructure;
+namespace Yiisoft\Hydrator\Attribute\Data;
+
+use Yiisoft\Hydrator\Attribute\Data\DataAttributeResolverInterface;
 
 /**
  * An interface for data attributes (allowed only in classes). Can be used to change the way data to be assigned is
