+++ src/AttributeHandling/ResolverFactory/AttributeResolverFactoryInterface.php
@@ -2,8 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Hydrator\AttributeInfrastructure;
+namespace Yiisoft\Hydrator\AttributeHandling\ResolverFactory;
 
+use Yiisoft\Hydrator\Attribute\Data\DataAttributeInterface;
+use Yiisoft\Hydrator\Attribute\Parameter\ParameterAttributeInterface;
 use Yiisoft\Hydrator\NonInstantiableException;
 
 interface AttributeResolverFactoryInterface
