+++ src/ObjectFactory/ReflectionObjectFactory.php
@@ -6,7 +6,7 @@ namespace Yiisoft\Hydrator\ObjectFactory;
 
 use ReflectionClass;
 use Yiisoft\Hydrator\NonInstantiableException;
-use Yiisoft\Hydrator\ObjectFactoryInterface;
+use Yiisoft\Hydrator\ObjectFactory\ObjectFactoryInterface;
 
 use function count;
