+++ src/TypeCaster/PhpNativeTypeCaster.php
@@ -9,8 +9,8 @@ use ReflectionType;
 use ReflectionUnionType;
 use Stringable;
 use Yiisoft\Hydrator\Result;
-use Yiisoft\Hydrator\TypeCastContext;
-use Yiisoft\Hydrator\TypeCasterInterface;
+use Yiisoft\Hydrator\TypeCaster\TypeCastContext;
+use Yiisoft\Hydrator\TypeCaster\TypeCasterInterface;
 use Yiisoft\Strings\NumericHelper;
 
 use function is_array;
