+++ src/TypeCaster/TypeCasterInterface.php
@@ -2,9 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Hydrator;
+namespace Yiisoft\Hydrator\TypeCaster;
 
 use ReflectionType;
+use Yiisoft\Hydrator\Result;
+use Yiisoft\Hydrator\TypeCaster\TypeCastContext;
 
 /**
  * Casts value to a type obtained from {@see ReflectionType} passed.
