+++ tests/AttributeHandling/ParameterAttributeResolveContextTest.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Hydrator\Tests;
+namespace Yiisoft\Hydrator\Tests\AttributeHandling;
 
 use PHPUnit\Framework\TestCase;
 use ReflectionClass;
-use Yiisoft\Hydrator\AttributeInfrastructure\ParameterAttributeResolveContext;
+use Yiisoft\Hydrator\AttributeHandling\ParameterAttributeResolveContext;
 use Yiisoft\Hydrator\Result;
 
 final class ParameterAttributeResolveContextTest extends TestCase
