+++ tests/Support/SkipTypeCaster.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Yiisoft\Hydrator\Tests\Support;
 
-use Yiisoft\Hydrator\TypeCastContext;
-use Yiisoft\Hydrator\TypeCasterInterface;
+use Yiisoft\Hydrator\TypeCaster\TypeCastContext;
+use Yiisoft\Hydrator\TypeCaster\TypeCasterInterface;
 use Yiisoft\Hydrator\Result;
 
 final class SkipTypeCaster implements TypeCasterInterface
