+++ tests/Support/String42TypeCaster.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Yiisoft\Hydrator\Tests\Support;
 
 use ReflectionNamedType;
-use Yiisoft\Hydrator\TypeCastContext;
-use Yiisoft\Hydrator\TypeCasterInterface;
+use Yiisoft\Hydrator\TypeCaster\TypeCastContext;
+use Yiisoft\Hydrator\TypeCaster\TypeCasterInterface;
 use Yiisoft\Hydrator\Result;
 
 final class String42TypeCaster implements TypeCasterInterface
